Rework plugin order + consolidation.
