+++ client/ayon_flame/plugins/publish/collect_batchgroup.py
@@ -1,5 +1,7 @@
 import pyblish.api
 
+from ayon_core.pipeline import PublishError
+
 from ayon_flame.otio import utils
 
 
@@ -17,23 +19,27 @@ class CollectBatchgroup(pyblish.api.InstancePlugin):
     def process(self, instance):
         creator_identifier = instance.data["creator_identifier"]
 
+        # Only interested in flame batchgroup.
         if creator_identifier != "io.ayon.creators.flame.batchgroup":
-            # only interested in flame batchgroup
+            self.log.debug("Not a batchgroup workfile, ignore.")
             return
 
-        # Validate that attach_to_task has required keys
-        required_keys = ["task_name", "task_type"]
+        # Validate attach_to_task settings.
         missing_keys = [
-            k for k in required_keys if k not in self.attach_to_task]
+            key for key in ("task_name", "task_type")
+            if key not in self.attach_to_task
+        ]
         if missing_keys:
-            raise ValueError(str(
+            raise ValueError(
+                "Invalid batchgroup workfile instance. "
                 "CollectBatchgroup.attach_to_task is missing "
                 f"required keys: {missing_keys}. "
                 "Please ensure settings are properly configured."
-            ))
+            )
 
-        # Update parent instance tasks with attach_to_task
-        self._update_parent_instance_tasks(instance)
+        # Ensure parent shot instance will
+        # create expected batchgroup task
+        self._update_parent_shot_instance_tasks(instance)
 
         # update shot related shared attributes
         parent_instance_id = instance.data["parent_instance_id"]
@@ -48,8 +54,9 @@ class CollectBatchgroup(pyblish.api.InstancePlugin):
             otio_timeline, instance.data["clip_index"]
         )
         if not otio_clip:
-            raise RuntimeError(
-                f"Could not retrieve otioClip for shot {instance}")
+            raise PublishError(
+                f"Could not retrieve otioClip for shot {instance}"
+            )
 
         instance.data["otioClip"] = otio_clip
 
@@ -65,21 +72,44 @@ class CollectBatchgroup(pyblish.api.InstancePlugin):
             "outputNodeProperties": self.output_node_properties,
         })
         self.log.info(
-            f"Collected batchgroup workfile products for shot {instance}")
+            f"Collected batchgroup workfile products for shot {instance}"
+        )
 
-    def _update_parent_instance_tasks(self, instance):
-        """Update parent instance tasks with attach_to_task if not present."""
+    def _update_parent_shot_instance_tasks(self, instance):
+        """ Update parent instance shot tasks with expected batchgroup task.
+        """
         parent_instance_id = instance.data["parent_instance_id"]
-        context = instance.context
-        for parent_instance in context:
-            if parent_instance.data["instance_id"] == parent_instance_id:
-                tasks = parent_instance.data.get("tasks", {})
-                attach_task_name = self.attach_to_task["task_name"]
-                new_task_data = {"type": self.attach_to_task["task_type"]}
-                if attach_task_name not in tasks:
-                    tasks[attach_task_name] = new_task_data
-                else:
-                    tasks[attach_task_name].update(new_task_data)
+        shot_tasks = parent_instance = None
 
-                parent_instance.data["tasks"] = tasks
+        # Find parent instance from context
+        for inst in instance.context:
+            if inst.data["instance_id"] == parent_instance_id:
+                parent_instance = inst
+                shot_tasks = inst.data.get("tasks", {})
                 break
+
+        else:
+            raise PublishError(
+                f"Cannot find parent shot instance: {parent_instance_id} "
+                f'for batchgroup instance: {instance.data["instance_id"]}'
+            )
+
+        # Batchgroup product get published under a specific task
+        # type that get created by the shot product. If expected batchgroup
+        # task is missing, warn and add it to the shot parent instance.
+        bg_task_name = self.attach_to_task["task_name"]
+        bg_task_type = self.attach_to_task["task_type"]
+        bg_task_data = {"type": bg_task_type}
+
+        if bg_task_name not in shot_tasks:
+            self.log.warning(
+                "Batchgroup settings expect a %s "
+                "(%s)task at shot level. Adding it "
+                "in the shot parent instance.",
+                bg_task_name, bg_task_type,
+            )
+            shot_tasks[bg_task_name] = bg_task_data
+        else:
+            shot_tasks[bg_task_name].update(bg_task_data)
+
+        parent_instance.data["tasks"] = shot_tasks
